--- environment.yml
@@ -1,11 +1,0 @@
-name: itslive_vortex
-channels:
-  - conda-forge
-dependencies:
-  - python=3.9.*
-  - ipyleaflet
-  - geopandas
-  - matplotlib-base
-  - pip
-  - pip:
-    - itslive
