+++ itslive/__init__.py
@@ -1,8 +1,8 @@
 from importlib.metadata import version
 
-from itslive import data_cube
+import itslive.velocity_cubes._cubes as velocity_cube
 
-__all__ = ["data_cube"]
+__all__ = ["velocity_cube", "velocity_pairs"]
 
 # this comes from the installed version not the editable source
 __version__ = version("itslive")
