--- itslive/data_cubes/__init__.py
@@ -1,2 +1,0 @@
-from ._data_cubes import * # imports all classes and functions in  __all__ into itslive.data_cubes namespace (see https://stackoverflow.com/questions/16509012/hide-external-modules-when-importing-a-module-e-g-regarding-code-completion )
-
