+++ itslive/dataviz/__init__.py
@@ -1,0 +1,3 @@
+from itslive.dataviz._viz import plot_terminal, plot_variable
+
+__all__ = ["plot_terminal", "[plot_variable"]
