+++ itslive/velocity_cubes/__init__.py
@@ -1,0 +1,19 @@
+from itslive.velocity_cubes._cubes import (export_csv, export_netcdf,
+                                           export_stdout, find, find_by_bbox,
+                                           find_by_point, find_by_polygon,
+                                           get_time_series, load_catalog,
+                                           plot_time_series_terminal)
+
+__all__ = [
+    "find",
+    "find_by_point",
+    "find_by_bbox",
+    "find_by_polygon",
+    "export_csv",
+    "get_time_series",
+    "load_catalog",
+    "export_netcdf",
+    "export_stdout",
+    "plot_time_series_terminal",
+    "plot_time_series",
+]
