+++ itslive/velocity_pairs/__init__.py
@@ -1,0 +1,3 @@
+from itslive.velocity_pairs._pairs import find
+
+__all__ = ["find"]
